Split out modules into KMP + JS
