+++ appJs/appJs/src/test/kotlin/com/gchristov/thecodinglove/MessengerTest.kt
@@ -7,7 +7,7 @@ class MessengerTest {
     @Test
     fun testGreet() {
         assertEquals(
-            expected = "Hello from multi-module Kotlin JS! Nested modules: ModuleA + ModuleB!",
+            expected = "Hello from multi-module Kotlin JS! Nested modules: ModuleA + KmpModuleB!",
             actual = Messenger().message()
         )
     }
