+++ appJs/module-a/src/main/kotlin/com/gchristov/thecodinglove/modulea/ModuleA.kt
@@ -1,0 +1,7 @@
+package com.gchristov.thecodinglove.modulea
+
+import com.gchristov.thecodinglove.kmpmoduleb.KmpModuleB
+
+class ModuleA {
+    fun name() = "ModuleA + " + KmpModuleB().name()
+}
