+++ appJs/module-a/src/test/kotlin/com/gchristov/thecodinglove/modulea/ModuleATest.kt
@@ -1,0 +1,14 @@
+package com.gchristov.thecodinglove.modulea
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class ModuleATest {
+    @Test
+    fun testGreet() {
+        assertEquals(
+            expected = "ModuleA + KmpModuleB",
+            actual = ModuleA().name()
+        )
+    }
+}
