--- appJs/modulea/src/main/kotlin/com/gchristov/thecodinglove/modulea/ModuleA.kt
@@ -1,7 +1,0 @@
-package com.gchristov.thecodinglove.modulea
-
-import com.gchristov.thecodinglove.moduleb.ModuleB
-
-class ModuleA {
-    fun name() = "ModuleA + " + ModuleB().name()
-}
