--- appJs/moduleb/src/main/kotlin/com/gchristov/thecodinglove/moduleb/ModuleB.kt
@@ -1,5 +1,0 @@
-package com.gchristov.thecodinglove.moduleb
-
-class ModuleB {
-    fun name() = "ModuleB"
-}
