--- appJs/moduleb/src/test/kotlin/com/gchristov/thecodinglove/moduleb/ModuleBTest.kt
@@ -1,14 +1,0 @@
-package com.gchristov.thecodinglove.moduleb
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ModuleBTest {
-    @Test
-    fun testGreet() {
-        assertEquals(
-            expected = "ModuleB",
-            actual = ModuleB().name()
-        )
-    }
-}
