+++ kmp/kmp-module-b/src/main/kotlin/com/gchristov/thecodinglove/kmpmoduleb/KmpModuleB.kt
@@ -1,0 +1,5 @@
+package com.gchristov.thecodinglove.kmpmoduleb
+
+class KmpModuleB {
+    fun name() = "KmpModuleB"
+}
