+++ kmp/kmp-module-b/src/test/kotlin/com/gchristov/thecodinglove/moduleb/ModuleBTest.kt
@@ -1,0 +1,15 @@
+package com.gchristov.thecodinglove.moduleb
+
+import com.gchristov.thecodinglove.kmpmoduleb.KmpModuleB
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class ModuleBTest {
+    @Test
+    fun testGreet() {
+        assertEquals(
+            expected = "KmpModuleB",
+            actual = KmpModuleB().name()
+        )
+    }
+}
